+++ warp-grpc/ChangeLog.md
@@ -1,8 +1,18 @@
 # Changelog for warp-grpc
 
-## Unreleased changes
+## 0.4.0.1
 
-## 0.2.0.0
+Add the `grpc-encoding` header to the response ([#20](https://github.com/haskell-grpc-native/http2-grpc-haskell/pull/20)).
+
+## 0.4.0.0
+
+Support for handlers in transformers stacks, and in monads which support [`unliftio`](http://hackage.haskell.org/package/unliftio).
+
+## 0.3.0.0
+
+Support for warp-3.3.x .
+
+## 0.2.0.0
 
 Support for different Protocol Buffers serialization libraries.